# openapi.json
{"openapi":"3.1.0","info":{"title":"Smart Home API","version":"v1"},"servers":[{"url":"http://localhost:8080","description":"Local"}],"tags":[{"name":"Devices","description":"Operations for managing devices"}],"paths":{"/api/users/{id}":{"get":{"tags":["user-controller"],"operationId":"get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/HomeUser"}}}}}},"put":{"tags":["user-controller"],"operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HomeUser"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/HomeUser"}}}}}},"delete":{"tags":["user-controller"],"operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}},"/api/schedules/{id}":{"get":{"tags":["schedule-controller"],"operationId":"get_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ScheduleView"}}}}}},"put":{"tags":["schedule-controller"],"operationId":"put","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScheduleDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ScheduleView"}}}}}},"delete":{"tags":["schedule-controller"],"operationId":"delete_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/rooms/{id}":{"get":{"tags":["room-controller"],"operationId":"get_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RoomView"}}}}}},"put":{"tags":["room-controller"],"operationId":"put_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoomDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RoomView"}}}}}},"delete":{"tags":["room-controller"],"operationId":"delete_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/devices/{id}":{"get":{"tags":["Devices"],"operationId":"get_3","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DeviceDTO"}}}}}},"put":{"tags":["Devices"],"operationId":"update_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeviceDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DeviceDTO"}}}}}},"delete":{"tags":["Devices"],"operationId":"delete_3","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}},"/api/users":{"get":{"tags":["user-controller"],"operationId":"list","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/HomeUser"}}}}}}},"post":{"tags":["user-controller"],"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HomeUser"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/HomeUser"}}}}}}},"/api/devices":{"get":{"tags":["Devices"],"operationId":"list_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DeviceDTO"}}}}}}},"post":{"tags":["Devices"],"operationId":"create_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeviceDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DeviceDTO"}}}}}}},"/api/devices/{id}/air-conditioner/auto-adjust":{"post":{"tags":["Devices"],"operationId":"autoAdjust","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AirConditionerDeviceDTO"}}}}}}},"/weather":{"get":{"tags":["weather-controller"],"operationId":"getWeather","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/WeatherResponse"}}}}}}},"/api/schedules":{"get":{"tags":["schedule-controller"],"operationId":"list_2","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ScheduleView"}}}}}}}},"/api/rooms":{"get":{"tags":["room-controller"],"operationId":"list_3","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RoomView"}}}}}}}},"/api/rooms/{id}/devices":{"get":{"tags":["room-controller"],"operationId":"devices","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"type":"string"},"uniqueItems":true}}}}}}},"/api/rooms/{id}/devices/{deviceId}":{"delete":{"tags":["room-controller"],"operationId":"removeDevice","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"deviceId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RoomView"}}}}}}}},"components":{"schemas":{"HomeUser":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"role":{"type":"string","enum":["OWNER","FAMILY_MEMBER","GUEST"]},"name":{"type":"string","maxLength":32,"minLength":2}},"required":["role"]},"ScheduleDto":{"type":"object","properties":{"targetType":{"type":"string","enum":["ROOM","DEVICE"]},"targetId":{"type":"string","minLength":1},"daysOfWeek":{"type":"array","items":{"type":"string","enum":["MONDAY","TUESDAY","WEDNESDAY","THURSDAY","FRIDAY","SATURDAY","SUNDAY"]},"maxItems":2147483647,"minItems":1,"uniqueItems":true},"startTime":{"type":"string"},"endTime":{"type":"string"},"action":{"type":"string","enum":["ON","OFF","TOGGLE"]},"timezone":{"type":"string","minLength":1}},"required":["action","daysOfWeek","endTime","startTime","targetId","targetType","timezone"]},"ScheduleView":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"targetType":{"type":"string","enum":["ROOM","DEVICE"]},"targetId":{"type":"string"},"daysOfWeek":{"type":"array","items":{"type":"string","enum":["MONDAY","TUESDAY","WEDNESDAY","THURSDAY","FRIDAY","SATURDAY","SUNDAY"]},"uniqueItems":true},"startTime":{"type":"string"},"endTime":{"type":"string"},"action":{"type":"string","enum":["ON","OFF","TOGGLE"]},"timezone":{"type":"string"}}},"RoomDto":{"type":"object","properties":{"name":{"type":"string","maxLength":50,"minLength":0},"description":{"type":"string"},"deviceIds":{"type":"array","items":{"type":"string"},"uniqueItems":true}},"required":["name"]},"RoomView":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"description":{"type":"string"},"deviceIds":{"type":"array","items":{"type":"string"},"uniqueItems":true}}},"DeviceDTO":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"deviceType":{"type":"string","enum":["UNSPECIFIED","SECURITY","MEDIA","AIR_CONDITIONER","VACUUM_CLEANER","LIGHTS"]},"name":{"type":"string","maxLength":32,"minLength":2},"roomId":{"type":"string"},"groupIds":{"type":"array","items":{"type":"string"}}},"required":["deviceType"]},"AirConditionerDeviceDTO":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"deviceType":{"type":"string","enum":["UNSPECIFIED","SECURITY","MEDIA","AIR_CONDITIONER","VACUUM_CLEANER","LIGHTS"]},"name":{"type":"string","maxLength":32,"minLength":2},"roomId":{"type":"string"},"groupIds":{"type":"array","items":{"type":"string"}},"temperature":{"type":"number","format":"double"},"humidityPercent":{"type":"number","format":"double"},"powerPercent":{"type":"number","format":"double"},"filterIsOn":{"type":"boolean"}},"required":["deviceType"]},"Main":{"type":"object","properties":{"temp":{"type":"number","format":"double"},"pressure":{"type":"integer","format":"int32"},"humidity":{"type":"integer","format":"int32"}}},"Weather":{"type":"object","properties":{"description":{"type":"string"}}},"WeatherResponse":{"type":"object","properties":{"main":{"$ref":"#/components/schemas/Main"},"weather":{"type":"array","items":{"$ref":"#/components/schemas/Weather"}},"name":{"type":"string"}}}}}}